requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-spectacular==0.27.1
anthropic==0.18.1
daphne==4.1.0
